third_party/WebKit/LayoutTests/http/tests/xmlhttprequest/navigation-should-abort-expected.txt

CONSOLE MESSAGE: line 16: PASS: Expected 'abort', got 'abort'.
If this text shows up, you've successfully navigated to 'navigation-target.html'.
